requirements.txt
# Core runtime dependencies
numpy
opencv-python-headless
ffmpeg-python
ultralytics
torch
torchvision
fastapi
uvicorn[standard]

# Utilities
pyyaml
python-dateutil
streamlit

# For packaging / CLI
click
